fix support for changing timezone
Unfortunately, python's os.stat() reports modification dates in local
time, which messes up synchronization with drive after moving to a
different time zone.

Fix this by reading the local timezone from the system using dateutil.tz
and use this for the time reported by os.stat(), as well as including it
in the modification time reported to Google Drive.